+++ lib/pusher-fake/cucumber.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-require "pusher-fake/support/cucumber"
-
-warn "[DEPRECATION] \"pusher-fake/cucumber\" is deprecated." \
-     "Please use \"pusher-fake/support/cucumber\" instead."